Split var expansion into norm-compliant blocks, minisplit pending
