+++ Makefile
@@ -6,7 +6,7 @@
 #    By: plouda <plouda@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/11/03 15:05:09 by plouda            #+#    #+#              #
-#    Updated: 2023/07/24 11:26:09 by plouda           ###   ########.fr        #
+#    Updated: 2023/07/24 12:27:47 by plouda           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -16,7 +16,8 @@ SRCS =	project/main.c \
 		project/lexer.c \
 		project/minisplit.c \
 		project/sanitizer.c \
-		project/sanitizer_helpers.c
+		project/sanitizer_helpers.c \
+		project/expand_vars.c
 
 BUILD_DIR = ./obj
 OBJS = $(addprefix $(BUILD_DIR)/, $(SRCS))
